move update-email-preference to general api folder
